requirements.txt
django==1.11
psycopg2==2.7.3.1
pyyaml==3.12
jinja2==2.4.0
pdfkit==0.6.1
urllib3==1.22